fix(ScrollView): Adds `manipulationModes` prop to enable gestures beh… (#1228)
Currently, the native ScrollView captures the pointer before any views inside the scroll view have a chance to respond to gestures. This breaks things like `SwipeableListView`, which uses a `PanResponder` behind the ScrollView to intercept X-dimension motions. This change adds a new prop to `ViewPropTypes` to support `ManipulationMode` on the UWP FrameworkElement. You must set the `ManipulationMode` of the view inside the `ScrollViewer` in order for it to respond to gestures without the pointer being captured by `ScrollViewer`.

Fixes #569
# Conflicts:
#	ReactWindows/ReactNative/UIManager/BaseViewManager.cs