Fix DependencyObjectData leak (#1561)
DependencyObjectExtensions maintains DependencyObjectData objects attached to each native view.
There is a ClearData cleanup method that's supposed to be called when a view is destroyed, but the current code doesn't do that.
As a result memory usage increases during the life time of an application.
The fix is to call ClearData from NativeViewHierarchyManager.DropView.

This exposes some cases where some asynchronous completion of an operation initiated by a view happens after the view is unmounted. Added some checks to prevent exceptions in one of these cases.
There may be other cases as well!

A more ideal solution would do the cleanup when the view reaches a reference count of 0, but that would require some major changes of the view lifetime, etc.